ticketing/tickets/src/routes/update.ts
import express, { Request, Response } from "express";
import { Ticket } from '../models/ticket';
import { body, param } from 'express-validator';
import { validateRequest, NotFoundError, requireAuth, NotAuthorizedError, BadRequestError } from "@anqtickets/common";
import { TicketUpdatedPublisher } from "../events/publishers/ticket-updated-publisher";
import { natsWrapper } from "../nats-wrapper";

const router = express.Router();

router.put(
    '/api/tickets/:id',
    requireAuth,
    [
        body('price')
            .isFloat({ gt: 0 })
            .withMessage('Price must be greater than 0')
    ],
    validateRequest,
    async (req: Request, res: Response) => {
        const { id } = req.params;
        const ticket = await Ticket.findById(id);
        if (!ticket) {
            throw new NotFoundError();
        }

        if (ticket.orderId) {
            throw new BadRequestError('Cannot edit a resrved ticket');
        }

        if (ticket.userId !== req.currentUser.id) {
            throw new NotAuthorizedError();
        }

        ticket.set({
            title: req.body.title,
            price: req.body.price,
        });

        try {
            await ticket.save();
        } catch(err) {
            throw new BadRequestError(err.message);
        }

        new TicketUpdatedPublisher(natsWrapper.client).publish({
            id: ticket.id,
            version: ticket.version,
            title: ticket.title,
            price: ticket.price,
            userId: ticket.userId,
        });
        
        return res.send(ticket);
    }
);

export { router as updateTicketRouter };